Extract common code from HTML and AMP shared modules
AMP is basically just a custom version of HTML. Therefore, the code for
the two exporters is very similar. For better reusability (e.g. when
implementing a Facebook Instant Article exporter), extract the code into
a `Common` module which then is reused by both exporters.
