+++ lib/article_json.rb
@@ -39,14 +39,14 @@ require_relative 'article_json/import/google_doc/html/embedded_slideshare_parser
 require_relative 'article_json/import/google_doc/html/embedded_tweet_parser'
 require_relative 'article_json/import/google_doc/html/parser'
 
+require_relative 'article_json/export/common/html/elements/shared/caption'
+require_relative 'article_json/export/common/html/elements/shared/float'
 require_relative 'article_json/export/common/html/elements/base'
 require_relative 'article_json/export/common/html/elements/text'
 require_relative 'article_json/export/common/html/elements/heading'
 require_relative 'article_json/export/common/html/elements/paragraph'
 require_relative 'article_json/export/common/html/elements/list'
 
-require_relative 'article_json/export/html/elements/shared/caption'
-require_relative 'article_json/export/html/elements/shared/float'
 require_relative 'article_json/export/html/elements/base'
 require_relative 'article_json/export/html/elements/text'
 require_relative 'article_json/export/html/elements/heading'
@@ -58,8 +58,6 @@ require_relative 'article_json/export/html/elements/quote'
 require_relative 'article_json/export/html/elements/embed'
 require_relative 'article_json/export/html/exporter'
 
-require_relative 'article_json/export/amp/elements/shared/caption'
-require_relative 'article_json/export/amp/elements/shared/float'
 require_relative 'article_json/export/amp/elements/base'
 require_relative 'article_json/export/amp/elements/text'
 require_relative 'article_json/export/amp/elements/paragraph'
