--- lib/article_json/export/html/elements/shared/float.rb
@@ -1,17 +1,0 @@
-module ArticleJSON
-  module Export
-    module HTML
-      module Elements
-        module Shared
-          module Float
-            # The element's floating class, if necessary
-            # @return [String]
-            def floating_class
-              "float-#{@element.float}" unless @element.float.nil?
-            end
-          end
-        end
-      end
-    end
-  end
-end
